テストコードのディレクトリの整理
