+++ MCI.Core.Tests/DateTimeHelperTests.cs
@@ -1,11 +1,6 @@
-namespace Miharu.Helpers
+namespace Miharu
 {
-    using System;
-    using System.Collections.Generic;
     using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-
     using Xunit;
 
     public class DateTimeHelperTests
