+++ MCI.Core.Tests/DayOfWeekExtensionsTests.cs
@@ -1,11 +1,7 @@
-namespace Miharu.Extensions
+namespace Miharu
 {
     using System;
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-
     using Xunit;
 
     public class DayOfWeekExtensionsTests
