+++ MCI.Core.Tests/PromiseTests.cs
@@ -1,10 +1,9 @@
-namespace Miharu.Monads
+namespace Miharu
 {
     using System;
     using System.Collections.Generic;
     using System.Text;
     using Xunit;
-    using Miharu;
 
     public class PromiseTests
     {
