+++ MCI.Core.Tests/StringExtensionsTests.cs
@@ -1,10 +1,6 @@
-namespace Miharu.Extensions
+namespace Miharu
 {
-    using System;
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
     using Xunit;
 
     public class StringExtensionsTests
